+++ cmake/onnxruntime_kernel_explorer.cmake
@@ -23,8 +23,7 @@ file(GLOB kernel_explorer_kernel_srcs CONFIGURE_DEPENDS "${KERNEL_EXPLORER_ROOT}
 
 onnxruntime_add_shared_library_module(kernel_explorer
   ${kernel_explorer_srcs}
-  ${kernel_explorer_kernel_srcs}
-  ${BERT_DIR}/util.cc)
+  ${kernel_explorer_kernel_srcs})
 set_target_properties(kernel_explorer PROPERTIES PREFIX "_")
 target_include_directories(kernel_explorer PUBLIC
   $<TARGET_PROPERTY:onnxruntime_pybind11_state,INCLUDE_DIRECTORIES>
