+++ onnxruntime/python/tools/kernel_explorer/device_array.h
@@ -6,7 +6,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
 #include "core/providers/rocm/rocm_common.h"
-#include "contrib_ops/rocm/bert/util.h"
+#include "core/providers/rocm/tunable/util.h"
 
 namespace py = pybind11;
