+++ onnxruntime/python/tools/kernel_explorer/kernel_explorer_interface.h
@@ -4,9 +4,10 @@
 #pragma once
 
 #include <hip/hip_runtime.h>
-#include "contrib_ops/rocm/bert/util.h"
+#include "core/providers/rocm/tunable/tunable.h"
+#include "core/providers/rocm/tunable/util.h"
 
-using onnxruntime::contrib::rocm::Timer;
+using onnxruntime::rocm::tunable::Timer;
 
 /// Wrapping around Op and TunableOp
 class IKernelExplorer {
