+++ onnxruntime/python/tools/kernel_explorer/kernels/gemm.cc
@@ -2,12 +2,16 @@
 // Licensed under the MIT License.
 
 #include "python/tools/kernel_explorer/kernels/gemm.h"
+
+#include <pybind11/pybind11.h>
+#include <type_traits>
+
+#include "core/providers/rocm/tunable/gemm_common.h"
 #include "python/tools/kernel_explorer/kernels/gemm_ck.h"
 #include "python/tools/kernel_explorer/kernels/gemm_rocblas.h"
 #include "python/tools/kernel_explorer/kernels/gemm_tunable.h"
 
-#include <type_traits>
-#include <pybind11/pybind11.h>
+using BlasOp = onnxruntime::rocm::tunable::blas::BlasOp;
 
 namespace py = pybind11;
